+++ clikt-mordant/src/commonTest/kotlin/com/github/ajalt/clikt/testing/utils.kt
@@ -1,0 +1,13 @@
+package com.github.ajalt.clikt.testing
+
+import com.github.ajalt.clikt.core.UsageError
+import com.github.ajalt.clikt.output.Localization
+import com.github.ajalt.clikt.output.ParameterFormatter
+
+val Throwable.formattedMessage: String?
+    get() = (this as? UsageError)?.formatMessage(
+        context?.localization ?: defaultLocalization,
+        ParameterFormatter.Plain
+    ) ?: message
+
+internal val defaultLocalization = object : Localization {}