+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/mpp/MppH.kt
@@ -1,9 +1,5 @@
 package com.github.ajalt.clikt.mpp
 
-internal val ANSI_CODE_RE = Regex("${"\u001B"}\\[[^m]*m")
-
-internal expect val String.graphemeLengthMpp: Int
-
 internal expect fun readEnvvar(key: String): String?
 
 internal expect fun isWindowsMpp(): Boolean