+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/parameters/transform/TransformContext.kt
@@ -2,9 +2,6 @@ package com.github.ajalt.clikt.parameters.transform
 
 import com.github.ajalt.clikt.core.Context
 import com.github.ajalt.clikt.core.UsageError
-import com.github.ajalt.clikt.core.terminal
-import com.github.ajalt.mordant.rendering.Theme
-import com.github.ajalt.mordant.terminal.Terminal
 
 interface TransformContext {
     /** The current context object */
@@ -18,12 +15,6 @@ interface TransformContext {
 fun TransformContext.message(message: String) {
     context.command.issueMessage(message)
 }
-
-/** The terminal for the current context */
-val TransformContext.terminal: Terminal get() = context.terminal
-
-/** The theme for the current context */
-val TransformContext.theme: Theme get() = terminal.theme
 
 data class HelpTransformContext(override val context: Context) : TransformContext {
     override fun fail(message: String): Nothing {